2022.11.7

+++ Services/CharacterService/CharacterService.cs
@@ -2,12 +2,13 @@ using dotnet_rpg.Dtos.Character;
 using AutoMapper;
 using dotnet_rpg.Data;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace dotnet_rpg.Services.CharacterService
 {
     public class CharacterService : ICharacterServices
     {
-         private static List<Character> characters = new List<Character>(){
+        private static List<Character> characters = new List<Character>(){
             new Character(),
             new Character{
                 Id = 1,
@@ -17,32 +18,41 @@ namespace dotnet_rpg.Services.CharacterService
 
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper,DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User
+        .FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
-             _context.Characters.Add(character);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+            _context.Characters.Add(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c))
+            .ToListAsync();
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             // return new ServiceResponse<List<GetCharacterDto>>
             // {
             //     Data = characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList()
             // };
             var response = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters  = await _context.Characters
-            .Where(c => c.User.Id == userId)
+            var dbCharacters = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
             .ToListAsync();
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return response;
@@ -55,7 +65,10 @@ namespace dotnet_rpg.Services.CharacterService
             // serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
             // return serviceResponse;
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+            .Include(c => c.Weapon)
+            .Include(c => c.Skills)
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -66,18 +79,29 @@ namespace dotnet_rpg.Services.CharacterService
             try
             {
                 //Character character = characters.FirstOrDefault(c => c.Id == updateCharacter.Id);
-                var character = await _context.Characters.FirstOrDefaultAsync(c=>c.Id == updateCharacter.Id);
+                var character = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
 
-                _mapper.Map(updateCharacter, character);
-                // character.Name = updateCharacter.Name;
-                // character.HitPoints = updateCharacter.HitPoints;
-                // character.Strength = updateCharacter.Strength;
-                // character.Defense = updateCharacter.Defense;
-                // character.Class = updateCharacter.Class;
+                if (character.User.Id == GetUserId())
+                {
+                    _mapper.Map(updateCharacter, character);
+                    // character.Name = updateCharacter.Name;
+                    // character.HitPoints = updateCharacter.HitPoints;
+                    // character.Strength = updateCharacter.Strength;
+                    // character.Defense = updateCharacter.Defense;
+                    // character.Class = updateCharacter.Class;
 
-                await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();
 
-                response.Data = _mapper.Map<GetCharacterDto>(character);
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Charactor not found.";
+                }
+
             }
             catch (Exception ex)
             {
@@ -90,14 +114,24 @@ namespace dotnet_rpg.Services.CharacterService
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
         {
-             ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();
+            ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                response.Data = _context.Characters.Select(c =>_mapper.Map<GetCharacterDto>(c)).ToList();
-                
+                Character character = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = _context.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
             }
             catch (Exception ex)
             {
@@ -106,6 +140,47 @@ namespace dotnet_rpg.Services.CharacterService
             }
             return response;
 
+        }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+            try
+            {
+                var character = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .FirstOrDefaultAsync(c => c.Id == newCharacterSkill.CharacterId &&
+                c.User.Id == GetUserId());
+
+                if (character == null)
+                {
+                    response.Success = false;
+                    response.Message = "Charactor not found.";
+                    return response;
+
+                }
+                var skill = await _context.Skills.FirstOrDefaultAsync(s =>s.Id == newCharacterSkill.SkillId);
+                 if (skill == null)
+                {
+                    response.Success = false;
+                    response.Message = "Skill not found.";
+                    return response;
+
+                }
+                character.Skills.Add(skill);
+                await _context.SaveChangesAsync();
+                response.Data = _mapper.Map<GetCharacterDto>(character);
+
+
+            }
+            catch (Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+
+            }
+            return response;
         }
     }
 }
